backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
arrow==1.3.0
asgiref==3.7.2
async-generator==1.10
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
azure-core==1.32.0
azure-storage-blob==12.24.1
billiard==4.2.1
binaryornot==0.4.4
blinker==1.9.0
boto3==1.26.159
botocore==1.29.159
Brotli==1.1.0
celery==5.4.0
certifi==2026.1.4
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.38.0
colorama==0.4.6
ConfigArgParse==1.7
constantly==23.10.4
cookiecutter==2.6.0
cron-descriptor==1.4.5
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
Django==4.1.9
django-celery==3.1.17
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-filter==2.4.0
django-js-asset==3.0.1
django-mptt==0.16.0
django-redis==5.4.0
django-sslserver==0.22
django-storages==1.13.2
django-timezone-field==7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
fastapi==0.129.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
frozenlist==1.5.0
geographiclib==2.0
geopy==2.3.0
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.4
importlib_metadata==8.7.1
incremental==24.7.2
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
locust==2.33.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy_extensions==1.1.0
numpy==2.4.2
openai==1.61.1
orjson==3.11.7
outcome==1.3.0.post0
packaging==26.0
pandas==3.0.0
pillow==11.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.7.0
pyOpenSSL==24.3.0
pypng==0.20220715.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.19.1
python-slugify==8.0.4
pytz==2023.3
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
qrcode==7.4.2
razorpay==2.0.0
redis==5.2.1
referencing==0.35.1
reportlab==4.4.10
requests==2.31.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.6.1
selenium==4.40.0
service-identity==24.2.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.2.4
starlette==0.52.1
text-unidecode==1.3
tqdm==4.67.1
trio==0.33.0
trio-typing==0.10.0
trio-websocket==0.12.2
Twisted==24.11.0
txaio==23.1.1
types-certifi==2021.10.8.3
types-python-dateutil==2.9.0.20241206
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.6.3
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.3.2
yarl==1.18.3
zipp==3.23.0
zope.event==5.0
zope.interface==7.2